mbls/requirements.txt
Flask
flask-restful
psycopg2-binary
redis